move _identical_files to tests
