+++ dpdispatcher/contexts/local_context.py
@@ -30,14 +30,6 @@ def _check_file_path(fname):
         os.makedirs(dirname, exist_ok=True)
 
 
-def _identical_files(fname0, fname1):
-    with open(fname0) as fp:
-        code0 = hashlib.sha1(fp.read().encode("utf-8")).hexdigest()
-    with open(fname1) as fp:
-        code1 = hashlib.sha1(fp.read().encode("utf-8")).hexdigest()
-    return code0 == code1
-
-
 class LocalContext(BaseContext):
     """Run jobs in the local server and remote directory.
